+++ java-desktop-util-chart/src/main/java/ec/util/chart/swing/JTimeSeriesChart.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2013 National Bank of Belgium
  *
- * Licensed under the EUPL, Version 1.1 or â€“ as soon they will be approved 
+ * Licensed under the EUPL, Version 1.1 or â€“ as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
  *
  * http://ec.europa.eu/idabc/eupl
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package ec.util.chart.swing;
@@ -633,7 +633,7 @@ public final class JTimeSeriesChart extends ATimeSeriesChart {
 
         @Override
         public boolean isObsHighlighted(int series, int item) {
-            return revealObs ^ obsHighlighter.apply(r.realIndexOf(series), item);
+            return (revealObs || dataset.getItemCount(series) == 1) ^ obsHighlighter.apply(r.realIndexOf(series), item);
         }
 
         @Override