+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix JDK25 file listing on blank filename [#504](https://github.com/nbbrd/java-desktop-util/issues/504)
+- Fix JGrid column header colors [#473](https://github.com/nbbrd/java-desktop-util/issues/473)
 
 ## [2.5.0] - 2025-07-10